+++ ref/factory.py
@@ -7,6 +7,8 @@ from ref.models import Subsidiary, PyouPyouUser
 
 test_tz = pytz.timezone("Europe/Paris")
 
+from itertools import permutations
+
 
 class PyouPyouUserFactory(factory.django.DjangoModelFactory):
     class Meta:
@@ -22,12 +24,20 @@ class PyouPyouUserFactory(factory.django.DjangoModelFactory):
     password = factory.Faker("password")
 
 
+def compute_subsidiary_code(name):
+    all_codes = Subsidiary.objects.values_list("code", flat=True)
+    for c in ["".join(perm).upper() for perm in permutations(name, 3)]:
+        if c.isalpha() and c not in all_codes:
+            return c
+    return ""  # will break at insertion as no combination is available
+
+
 class SubsidiaryFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = "ref.Subsidiary"
 
     name = factory.Faker("company")
-    code = factory.LazyAttribute(lambda n: n.name[0:3].upper())
+    code = factory.LazyAttribute(lambda n: compute_subsidiary_code(n.name))
 
 
 class ConsultantFactory(factory.django.DjangoModelFactory):