+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add others process on process detail for a candidate
 - Add global subsidiary filter
 - Add goal to new interview mail notification
+- Fix a bug where non-unique subsidiary codes could be created by its factory
 
 # v1.16.0 (2022-11-14)
 - More refined factories